- Workarounds for badly ordered result sets.
DatabaseElementList does not represent a list of elements but rather an ordered multiset of elements, ordered by schema id. This means that results sometimes appear in the wrong order. We workaround the problem by relying on the order of the schema elements to determine the order of the data elements.

- Tweaked map visualisation. Markers are now placed at intervals of 1/10th of a second.